// frontend/packages/studio/workspace/entry-adapter/src/pages/falcon/components/gridList.d.ts
export declare const GridList: ({ children, averageItemWidth, gap, className, onResize, ...resetProps }: {
    [x: string]: any;
    children: any;
    averageItemWidth?: number | undefined;
    gap?: number | undefined;
    className: any;
    onResize: any;
}) => import("react").JSX.Element;
export declare const GridItem: ({ children, disabled, className, ...restProps }: {
    [x: string]: any;
    children: any;
    disabled?: boolean | undefined;
    className: any;
}) => import("react").JSX.Element;